requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
aiohttp==3.8.4
yt-dlp==2023.3.4
gTTS==2.3.1
transformers==4.28.1
torch==1.13.1
diffusers==0.16.1
Pillow==9.5.0
opencv-python-headless==4.7.0.72
youtube_transcript_api==0.5.0
numpy==1.24.3
requests==2.30.0

